requirement.txt
django==4.1
django-cors-headers==4.3.0
djangorestframework==3.14.0
django-crontab==0.7.1
passlib==1.7.4
pyjwt==2.8.0
PyMysql==1.1.0
matplotlib==3.7.3
requests==2.31.0
tqdm==4.66.1
